pyproject.toml
[project]
name = "full-stack-fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faker>=37.6.0",
    "pytest>=8.4.1",
    "pytest-playwright>=0.7.0",
    "requests (>=2.32.5,<3.0.0)",
    "allure-pytest (>=2.15.0,<3.0.0)",
]
